+++ src/lib/Components/Consignments/Screens/Welcome.tsx
@@ -1,20 +1,74 @@
 import React from "react"
-import { Dimensions, NavigatorIOS, Route, ScrollView, Text, View, ViewProperties } from "react-native"
-
-import { Router } from "lib/utils/router"
-import SwitchBoard from "../../../NativeModules/SwitchBoard"
+import { Dimensions, Image, NavigatorIOS, Route, ScrollView, ViewProperties } from "react-native"
 
 import { Schema, screenTrack } from "lib/utils/track"
-import Circle from "../Components/CircleImage"
+import CloseButton from "../Components/CloseButton"
 import ConsignmentBG from "../Components/ConsignmentBG"
 import { Button } from "../Components/FormElements"
-import { BodyText as P, LargeHeadline, SmallPrint } from "../Typography"
+import { LargeHeadline } from "../Typography"
 import Overview from "./Overview"
+
+import { Fonts } from "lib/data/fonts"
+import styled from "styled-components/native"
 
 interface Props extends ViewProperties {
   navigator: NavigatorIOS
   route: Route
 }
+
+const isPad = Dimensions.get("window").width > 700
+
+const VerticalLayout = styled.View`
+  flex: 1;
+  flex-direction: column;
+  min-height: 400;
+  max-height: 800;
+  align-items: ${isPad ? "center" : "stretch"};
+`
+const Listpad = styled.View`
+  align-items: center;
+  padding-top: 80;
+`
+const Listphone = styled.View`margin-bottom: 100;`
+
+const HorizontalLayout = styled.View`
+  flex: 1;
+  flex-direction: row;
+  margin-left: ${isPad ? 0 : 20};
+  margin-top: 50;
+  margin-bottom: ${isPad ? 0 : 50};
+`
+
+const Title = styled.Text`
+  text-align: center;
+  font-size: ${isPad ? 46 : 16};
+  line-height: ${isPad ? 58 : 32};
+  width: ${isPad ? 760 : 300};
+  margin-top: ${isPad ? 80 : 35};
+  font-family: ${Fonts.AvantGardeRegular};
+  align-self: center;
+  color: white;
+`
+
+const SmallHeadline = styled(LargeHeadline)`
+  font-size: ${isPad ? 34 : 20};
+  color: white;
+  ${isPad ? "width: 540" : "max-width: 280"};
+  min-height: 60;
+  text-align: left;
+`
+
+const Icon = styled(Image)`
+  resize-mode: contain;
+  width: 40;
+`
+
+const ButtonsView = styled.View`
+  flex: 1;
+  align-items: center;
+`
+
+const List = isPad ? Listpad : Listphone
 
 @screenTrack({
   context_screen: Schema.PageNames.ConsignmentsWelcome,
@@ -23,63 +77,35 @@ interface Props extends ViewProperties {
 export default class Welcome extends React.Component<Props, null> {
   goTapped = () => this.props.navigator.push({ component: Overview })
 
-  // It's not optimal to dismiss the modal, but otherwise we get into all sorts of tricky states
-  privacyPolicyTapped = () =>
-    SwitchBoard.dismissModalViewController(this) &&
-    SwitchBoard.presentNavigationViewController(this, Router.PrivacyPage)
-
-  TOSTapped = () =>
-    SwitchBoard.dismissModalViewController(this) &&
-    SwitchBoard.presentNavigationViewController(this, Router.TermsOfService)
-
   render() {
-    const isPad = Dimensions.get("window").width > 700
-
-    const TOS = () =>
-      <Text style={{ textDecorationStyle: "solid", textDecorationLine: "underline" }} onPress={this.TOSTapped}>
-        Terms of Sevice
-      </Text>
-
-    const Privacy = () =>
-      <Text
-        style={{ textDecorationStyle: "solid", textDecorationLine: "underline" }}
-        onPress={this.privacyPolicyTapped}
-      >
-        Privacy Policy
-      </Text>
-
     return (
-      <ConsignmentBG showCloseButton>
+      <ConsignmentBG>
         <ScrollView style={{ flex: 1 }} centerContent>
-          <View
-            style={{
-              marginTop: 40,
-              alignItems: "center",
-              alignSelf: "center",
-              width: "100%",
-              maxWidth: 540,
-            }}
-          >
-            <LargeHeadline>Sell works from your collection through our partner network</LargeHeadline>
-
-            <View style={{ width: 300, alignItems: "center", marginVertical: isPad ? 60 : 20 }}>
-              <Circle source={require("../../../../../images/consignments/email.png")} />
-
-              <P>Receive offers from partner galleries and auction houses</P>
-              <Circle
-                source={require("../../../../../images/consignments/hammer.png")}
-                style={{ marginTop: isPad ? 30 : 0 }}
-              />
-
-              <P>Get your work placed in an upcoming sale</P>
-            </View>
-
-            <Button text="GET STARTED" onPress={this.goTapped} />
-
-            <SmallPrint style={{ width: 360, marginTop: 20 }}>
-              By submitting works to Artsy’s consignment program you agree to our {TOS()} and {Privacy()}
-            </SmallPrint>
-          </View>
+          <VerticalLayout>
+            <Title>SELL WORKS FROM YOUR COLLECTION</Title>
+            <List>
+              <HorizontalLayout>
+                <Icon source={require("../../../../../images/find.png")} />
+                <SmallHeadline>Take a few photos and submit details about the work</SmallHeadline>
+              </HorizontalLayout>
+              <HorizontalLayout>
+                <Icon source={require("../../../../../images/contact.png")} />
+                <SmallHeadline>Get the offers from galleries and auction houses</SmallHeadline>
+              </HorizontalLayout>
+              <HorizontalLayout>
+                <Icon source={require("../../../../../images/message.png")} />
+                <SmallHeadline>Have your work placed in a gallery or upcoming sale</SmallHeadline>
+              </HorizontalLayout>
+              <HorizontalLayout>
+                <Icon source={require("../../../../../images/pay.png")} />
+                <SmallHeadline>Receive payment once the work sells</SmallHeadline>
+              </HorizontalLayout>
+              <ButtonsView>
+                <Button text="GET STARTED" onPress={this.goTapped} />
+                <CloseButton />
+              </ButtonsView>
+            </List>
+          </VerticalLayout>
         </ScrollView>
       </ConsignmentBG>
     )